chore(kitList): refactoring code
